+++ api/src/opentrons/legacy_api/protocols/__init__.py
@@ -1,3 +1,6 @@
+""" Functions for executing protocols targeting v1
+"""
+
 from . import execute_v1, execute_v3
 from opentrons.protocol_api.execute import (
     get_protocol_schema_version, validate_protocol)
